pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tektonit"
version = "0.1.0"
description = "Autonomous agent that generates unit tests for Tekton Tasks, Pipelines, and StepActions"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
    "jinja2>=3.1",
    "jsonschema>=4.20",
    "gitpython>=3.1",
    "google-genai>=1.0",
    "PyGithub>=2.1",
    "tenacity>=8.2",
    "prometheus-client>=0.19",
    "python-json-logger>=2.0",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.42",
]
openai = [
    "openai>=1.50",
]
all = [
    "tektonit[anthropic,openai]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.4",
]

[project.scripts]
tektonit = "tektonit.cli:main"
tektonit-monitor = "tektonit.monitor:main"

[tool.setuptools.packages.find]
include = ["tektonit*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
